F-84: Remove osm plugin, add flutter map
