+++ client/pubspec.yaml
@@ -44,7 +44,9 @@ dependencies:
   riverpod_annotation: ^2.3.5
   unicons: ^3.0.0
   flutter_launcher_icons: ^0.14.1
-  flutter_osm_plugin: ^1.3.5
+  flutter_map: ^7.0.2
+  latlong2: ^0.9.1
+  cached_network_image: ^3.4.1
 
 dev_dependencies:
   flutter_test:
